requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML
openai>=1.10.0
anthropic>=0.18.1
pillow>=10.3.0

# Music API
spotipy>=2.23.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Optional: For local ML models
# torch==2.1.2
# transformers==4.37.0
# clip-by-openai==1.0
